Cache block boundaries in the JIT compiler.
This prevents continual recompilation for non-self-modifying cases.

Quite a few subtleties involving entering and returning from interrupts,
where it's important to avoid unneccessary recompilation and creation of
false block boundaries.

CLOCKSP performance is now to a combined average of 3175Mhz, a rapid jump
towards the historical high of about 6Ghz.