Use a fault-based approach for indirect reads / writes hitting registers.
This avoids tons of inline JIT code to check for register addresses after every
indirect read and write.

Instead, new "indirect" mappings are hit by the indirect instructions which hit
a page fault if they touch registers. This is very uncommon so we take a slow
but compatible uncommon case for a fast common case.

And also the SIGSEGV handler is back! Equal parts fun and horror.