Treat countdown checks as separate opcode objects with an internal opcode.
This fixes an issue where self-modified code at the start of a block was
getting tagged in a way that made it look like a new block.

This fix passes tests but seems to break kind of everything in real code.
More test cases to follow to fix the testing gap.