Declare and enforce invariant on timers vs. instruction boundary.
Now, timers at the instruction boundary will always fire before
the next instruction starts.
Previously it was a bit of a mixed bag, including differences
between interp vs. the other modes. It's unclear if this is the
root of timing problems with replay mode or not.

This change also tries to tidy and simplify the interp logic for
handling timers firing.