Accurately calculate the timer for vsync events.
Still a few asserts that fire from time to time so the printf's
are left in to get to the bottom of it.

Elite looks perfect.
Firetrack looks perfect.
Revs is a mess (needs interlace support?!)
Castle Quest loading screen looks perfect.
Tricky's Scramble is a hot mess.
Uridium looks perfect.