Fix building with clang on Linux.
(I used clang version 6.0.0-1ubuntu2 (tags/RELEASE_600/final).)

The assembler code uses "jmp 0" (etc.) in various places, to produce a
jump with a 4-byte offset. clang's assembler doesn't accept this syntax,
but you can force it to generate a 4-byte offset by jumping to an undefined
symbol. The assembler can't prove the label is close enough for a 1-byte
offset, so it generates a 4-byte one.

Then add a suitable C function, so it links. (The C function does an int 3,
so if an un-fixed-up branch is ever encountered, something obvious will
happen.)

This all seems to work just the same with gcc.