Don't bail upon double density 1770 commands.
This doesn't add support for MFM, but it does something sensible
instead of bailing.

This allows Advanced Disc Inspector (ADI) to copy empty tracks under
emulation. Upon seeing an empty track in FM mode, it tries to read
the track in MFM mode. We ensure it doesn't see anything in MFM
mode either.