Implement dynamic opcodes in the JIT compiler.
Unfortunately, it is disabled at this time. Despite all the tests
passing, it is very unstable.

Dyanmic opcodes are used when a code location is repeatadly
recompiled with different opcodes (or different operands where we
don't support that particular dynamic operand). The dynamic opcode
uses inturbo to interpret a single opcode quickly, without the
need for a recompile.