Change the x64 JIT model so 6502 X and Y don't include host pointer.
The old model was an attempt at speed, but its time seems to be
past.

Not including the host pointer in the X and Y registers simplifies
a few things.

There don't appear to be any performance regressions; if anything,
it's a shade faster. Not having to zero-extend the X and Y registers
saves a few bytes here are there, so perhaps that's decisive.