Tidy up JIT segment handling.
This prefers to hit K_BBC_MEM_READ_IND_ADDR in all cases where
that's possible, i.e. it's known that the address is plain RAM, i.e.
< 0x8000 in the beeb's memory map.

Doesn't seem to make a performance difference on my machine, but
there has been a case or two of L1 TLB thrashing in the past.