Change initialization of the dummy raster latch on new CRTC frame.
Initialize it the same way as if we hit R7: copy the "is odd interlace
frame" bit.
Normally this will have no effect as R7 will be hit later in the
frame anyway.
But for very unusual setups, such as MODE7-75, it'll hit.
This change fixes the last difference of MODE7-75 with real
hardware: a mysterious extra black scanline about a third of the
way down the screen is now correctly present.