Use a much shorter sub-based countdown sequence where possible.
If the code block replaces N,Z flags before using them, the
countdown sequence is free to trash N,Z flags. In this case, we
can use sub / jc as opposed to the much longer lea / bt / jc.

Seems to be good for 12.1GHz -> 12.3GHz on the RTRACE rocket.

CLOCKSP 17.4GHz combined, 22GHz Integer FOR loop. Not sure what
it was before, but this is the first time I remember seeing >20GHz
on any of the CLOCKSP sub-tests.