Reload mutable disc images upon Alt+R power-on reset.
A convenience asked for a few times by retro developers.

This also changes behavior for attempts to load FSD, SCP, and
similar files with -mutable. Previously they were converted to
HFE and now they are not (and are made not writeable in the beeb).
This can be revisited if necessary, but we'd need to decide how to
behave with Alt+R in such a case.

This chance can also be considered the start of a tidy-up. Creating
the disc object is now separated out from loading in the image file,
and the format conversion code is now in a separate function (not
publicly exposed yet).