Fix Ctrl-C in the Windows console.
It was only working once then it exited the process!
We now use a proper Windows API for trapping Ctrl-C and not signal,
which had SysV semantics on Windows.
Also tidy up the POSIX code -- use sigaction() instead of signal()
to get a guarantee that the SIGINT handler isn't reset.