Re-paint the screen when debugger blocks on user input.
This gives the expected visual results when debugging why your
latest beeb demo isn't rendering properly :)

This commit also introduces the option -opt video:always-clear,
which may be needed for the best results. It blanks the render
canvas after every vsync-induced paint. This means that the progess
of the video beam can be tracked much more easily, because we then
render on top of black instead of rendering on top of the previous
frame.