Land disabled support for branching around within compile blocks.
This is called "extended blocks".

I had hoped this would be faster due to better L1 icache utilization
and less jmp'ing around.

It seems to be slower though, which is surprising. It certainly
needs more investigating, but initial results were:

- Citadel attract. 7.3GHz normal, 6.7GHz extended blocks.
- RTRACE rocket. 8.2GHz in both cases.
- Starship Command attract. 5.3GHz normal, 4.8GHz extended blocks.

I'm keeping the code for future experimentation. It'll be
interesting to see if the story is similar on ARM64 or not.