1770: delay a little bit (32us) before ending a command.
This fixes a strange race condition hit with Watford DDFS 1.54T.
The symptom is that a write to a write protected disc appears to
succeed (despite not doing so).
Before this change, the command "write multiple sectors" aborted
immediately (NMI after 0us) and this hit a race condition inside
the Watford ROM.
Now we behave a bit more like a real 1770 and take a bit of time
to do anything, including fail a command immediately.