Fix assert regression in debugger while in fast mode.
The new code to sync the current framebuffer to the screen causes
asserts when in fast mode, because fast mode uses "inactive rendering"
for most of its frame, and you can't just advance CRTC time in
that state.

Changed so the general CRTC advance call handles "inactive rendering"
state.

Also noticed that this assert could be triggered since a long time
by the game "Pharoah's Curse", which performs the unusual step of
faking a light pen pulse, which is an unusual code path.