Fix an array of regression in JIT corner cases.
The interaction between metadata management and the optimizer got
broken some time recently.
It most affected the loading of protected tapes, which like to use
undocumented opcodes, causing a JIT bail-out to the interpreter.
Android Attack would assert() and Repton would fail to load.

And improve test coverage of the interaction.