Make a demo of self-modification invalidation on ARM64.
It just writes 0x00000000 but this is enough to demo what is
presumably the icache / dcache incoherency on ARM.
In non-debug, the test crashes because the self-invalidation
did nothing, i.e. the 0x00000000 did not execute.
In debug, the icache appears to get cycled hard enough (by
calling into C) that the 0x00000000 opcode executes and we
get illegal opcode.