Fix value -> ZPG stores on x64.
The sequence used was too long, and wrote to the "write indirect"
mapping. We can use a short 1-byte index sequence and write to the
"read indirect" mapping, which is fine for zero page accesses.

Surprisingly, this is +10% faster on the rockets.ssd RTRACE test.
It's not clear why. I don't think it's L1 icache usage. Perhaps it
is related to memory aliasing. By avoiding using a memory alias
for the zero page RMW cycles, we'll allow the memory subsystem to
do its thing better?