Get the JIT starting to run on macOS ARM64.
This fixes a raft of issues with hosting the JIT mapping in a
64-bit location (i.e. >4GB).
Carefully placed trampolines are used to bounce out of JIT and
into debug / interp, to avoid trying to jump to a delta that is
too great for the ARM64 architecture.
test.rom now gets a little way, and does some actual compile / run
of blocks, before hitting bus error.