Add MacOS keyboard support.
Has the quirk that left shift is used for BBC CAPS LOCK, because MacOS
has dubious caps lock reporting. Right shift is used for BBC SHIFT.
The MacOS keyboard on my MacBook Air has a general lack of keys so some
BBC keys may be tricky to hit.
e.g. the End key on modern keyboards is mapped to BBC COPY, and to hit
that on the MacBook Air, it's fn + right arrow.
"option" serves the function of Alt, i.e. option + f for fast mode.