Fix stale code cleanup on ARM64.
Was crashing due to not mixing in the JIT pointer base.

Now we can see the effect of clearing up the faulting invalidations.
Zalaga starfield went from 800MHz to 5GHz.

Still some work to do on Galaforce and others, which seem to still
be taking faults after the stale code cleanup.