x64: Re-implement the "lea -> sub" countdown optimization.
On reflection, it's possible to use "sub" for the countdown
check after all, in many conditions.
If the opcode immediately after the countdown sets NZ flags, then
it's ok to trash the NZ flags in the countdown check. We can do this
because even if the countdown fires and an IRQ is raised, the opcode
immediately after the countdown must execute before the IRQ is
seen and acted upon.

This is great for performance. On my i7-1165G7,

CLOCKSP                  : 19GHz   -> 20.2GHz
RTRACE                   : 12.4GHz -> 12.9GHz
Galaforce attract        : 4.7GHz  -> 5.1GHz
Chuckie Egg attract      : 7.4GHz  -> 8.3GHz
Citadel attract          : 6.9GHz  -> 7.5GHz