Add optimization to collapse wait loops into constant code.
A lot of games wait for an event by spinning in a loop that waits
for a memory value to change.
In the JIT, we know the memory value can't change until some external
event fires and triggers an IRQ. So we can collapse the loop into
a constant sequence.

Speedups range from nothing to significant.
Attract screens:
Citadel          7.0GHz -> 9.6GHz (!)
Thrust           5.8GHz -> 6.1GHz
Meteors          5.9GHz -> 6.4GHz
Palace of Magic  5.7GHz -> 9.0GHz (!)
Frak             2.6GHz -> 3.2GHz

One disappointment is that the replays of Exile / Castle Quest
did not seem to speed up at all, even though both these games
use this type of loop.

Needs to be added for ARM64.