Add framework for JIT encoded callbacks.
This is where a simple hardware register reference can be
respresented by a JIT opcode sequence, instead of the extremely
expensive operation of bouncing out of JIT and in to a C function.

To get the ball rolling, reads of $FE80 (the Intel 8271 FDC status
register) go through the new scheme.

A repeated $FE80 read loop goes from 390MHz to 4.2GHz.
A more practical test case, waiting for a read byte with the drive
spinning, goes from 310MHz to 810MHz. (It remains a little slow
because the spinning drive generates a timer event every 64us.)

Possible Tarzan attract mode regression from 780MHz -> 720MHz; it's
hard to suggest how!