Use a JIT encoded callback for ROMSEL writes.
It's common for games that use user VIA timers to trigger an
unidentified interrupt (as far as MOS is concerned). This has a
knock-on effect of a bit of ROM paging to see if any ROM wants to
handle it. Often, the paging requests are for the currently
selected ROM so they are fast and benefit from further speedup.

Barbarian attract, x64, goes from 10GHz to 10.4GHz.