Add (disabled, not working) port of full frame video fast mode.
Currently, video fast mode jumps to both VSYNC raise and lower. It
should be possible to only stop at VSYNC lower and avoid bouncing
out of JIT due to the video timer so much.

This is a port of an old patch. It doesn't yet work due to an
assert crash if you enable it.